+++ src/composition-root.ts
@@ -12,15 +12,8 @@ import { PostsQueryRepository } from './query-reositories/posts-query-repository
 import { UsersQueryRepository } from './query-reositories/users-query-repository'
 import { EmailAdapter } from './adapters/email-adapter'
 import { JwtService } from './application/jwt-service'
-import { AuthService } from './domain/auth-service'
 import { EmailManager } from './managers/email-manager'
-import { SessionsService } from './domain/sessions-service'
 import { AuthController } from './controllers/auth-controller'
-import { BlogsService } from './domain/blogs-service'
-import { CommentsService } from './domain/comments-service'
-import { PostsService } from './domain/posts-service'
-import { TestService } from './domain/test-service'
-import { UsersService } from './domain/users-service'
 import { BlogsController } from './controllers/blogs-controller'
 import { CommentsController } from './controllers/comments-controller'
 import { PostsController } from './controllers/posts-controller'
@@ -30,6 +23,13 @@ import { UsersController } from './controllers/users-controller'
 import { CommentsLikesRepository } from './repositories/comments-likes-repository'
 import { Container } from 'inversify'
 import { PostsLikesRepository } from './repositories/posts-likes-repository'
+import {AuthService} from "./services/auth-service";
+import {BlogsService} from "./services/blogs-service";
+import {CommentsService} from "./services/comments-service";
+import {PostsService} from "./services/posts-service";
+import {SessionsService} from "./services/sessions-service";
+import {TestService} from "./services/test-service";
+import {UsersService} from "./services/users-service";
 
 export const container = new Container()
 container.bind(BlogsRepository).to(BlogsRepository)
