+++ src/controllers/auth-controller.ts
@@ -7,15 +7,15 @@ import {
     RegistrationResendType
 } from '../models/auth-models'
 import { Request, Response } from 'express'
-import { AuthService } from '../domain/auth-service'
 import { HTTP_STATUSES } from '../constats/status'
 import { JwtService } from '../application/jwt-service'
 import { v4 as uuidv4 } from 'uuid'
-import { SessionsService } from '../domain/sessions-service'
 import { settings } from '../settings/settings'
 import { checkAuthorizationAndGetPayload } from './helper'
 import { UsersTypeInput } from '../models/users-models'
 import { inject, injectable } from 'inversify'
+import {AuthService} from "../services/auth-service";
+import {SessionsService} from "../services/sessions-service";
 
 const SECURE_COOKIE_MODE = settings.SECURE_COOKIE_MODE == 'true'
