+++ src/controllers/comments-controller.ts
@@ -6,13 +6,13 @@ import {
     CommentsTypeOutput
 } from '../models/comments-models'
 import { Response } from 'express'
-import { CommentsService } from '../domain/comments-service'
 import { HTTP_STATUSES } from '../constats/status'
 import { QueryComments } from '../models/query-models'
 import { PaginatedType } from '../models/main-models'
 import { CommentsQueryRepository } from '../query-reositories/comments-query-repository'
 import { LikesType } from '../models/likes-models'
 import { inject, injectable } from 'inversify'
+import {CommentsService} from "../services/comments-service";
 
 @injectable()
 export class CommentsController {
