+++ src/controllers/security-controller.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from 'express'
 import { checkAuthorizationAndGetPayload } from './helper'
 import { HTTP_STATUSES } from '../constats/status'
-import { SessionsService } from '../domain/sessions-service'
 import { RequestWithParams } from '../models/types'
 import { DeviceIdParams } from '../models/auth-models'
 import { inject, injectable } from 'inversify'
+import {SessionsService} from "../services/sessions-service";
 
 @injectable()
 export class SecurityController {
