+++ src/query-reositories/users-query-repository.ts
@@ -2,7 +2,7 @@ import { QueryUsers } from '../models/query-models'
 import { PaginatedType } from '../models/main-models'
 import { UsersTypeOutput } from '../models/users-models'
 import { getPaginatedType, makeDirectionToNumber } from './helper'
-import { UsersModel } from '../repositories/db'
+import { UsersModel } from '../repositories/db/db'
 import { injectable } from 'inversify'
 
 @injectable()
