+++ src/repositories/comments-repository.ts
@@ -1,5 +1,5 @@
 import { CommentsTypeInput, CommentsTypeOutput, CommentsTypeToDB } from '../models/comments-models'
-import { CommentsModel } from './db'
+import { CommentsModel } from './db/db'
 import { ObjectId } from 'mongodb'
 import { LikesInfoType, StatusLikeType } from '../models/likes-models'
 import { injectable } from 'inversify'
