+++ src/services/sessions-service.ts
@@ -1,19 +1,21 @@
-import jwt from 'jsonwebtoken'
-import { settings } from '../settings/settings'
-import { TokenPayloadType } from '../models/token-models'
-import { SessionsRepository } from '../repositories/sessions-repository'
-import { HTTP_STATUSES } from '../constats/status'
-import { SessionOutputType, SessionTypeToDB } from '../models/session-models'
-import { HttpTypes } from '../models/status-models'
-import { inject, injectable } from 'inversify'
+import {inject, injectable} from "inversify";
+import {SessionsRepository} from "../repositories/sessions-repository";
+import {SessionOutputType, SessionTypeToDB} from "../models/session-models";
+import jwt from "jsonwebtoken";
+import {settings} from "../settings/settings";
+import {TokenPayloadType} from "../models/token-models";
+import {HttpTypes} from "../models/status-models";
+import {HTTP_STATUSES} from "../constats/status";
 
 @injectable()
 export class SessionsService {
-    constructor(@inject(SessionsRepository) protected sessionsRepository: SessionsRepository) {}
+    constructor(@inject(SessionsRepository) protected sessionsRepository: SessionsRepository) {
+    }
 
     async findAllActiveSession(userId: string): Promise<SessionOutputType[] | null> {
         return await this.sessionsRepository.findAllActiveSession(userId)
     }
+
     async saveSession(token: string, ip: string, deviceName: string): Promise<void> {
         const result = jwt.verify(token, settings.JWT_REFRESH_SECRET) as TokenPayloadType
         const issueAt = new Date(result.iat * 1000).toISOString()
@@ -24,6 +26,7 @@ export class SessionsService {
         const session: SessionTypeToDB = new SessionTypeToDB(issueAt, deviceId, deviceName, ip, userId, expireAt)
         await this.sessionsRepository.saveSession(session)
     }
+
     async updateSession(token: string): Promise<void> {
         const result = jwt.verify(token, settings.JWT_REFRESH_SECRET) as TokenPayloadType
         const issueAt = new Date(result.iat * 1000).toISOString()
@@ -31,6 +34,7 @@ export class SessionsService {
         const deviceId = result.deviceId
         await this.sessionsRepository.updateSession(issueAt, expireAt, deviceId)
     }
+
     async isActiveSession(deviceId: string, iat: string): Promise<boolean> {
         const result = await this.sessionsRepository.findSession(iat)
         if (!result) {
@@ -38,10 +42,12 @@ export class SessionsService {
         }
         return result.deviceId === deviceId
     }
+
     async deleteSession(iat: number): Promise<boolean> {
         const issueAt = new Date(iat * 1000).toISOString()
         return await this.sessionsRepository.deleteSession(issueAt)
     }
+
     async deleteSessionByDeviceId(deviceId: string, userId: string): Promise<boolean | HttpTypes> {
         const device = await this.sessionsRepository.findSessionByDeviceId(deviceId)
 
@@ -54,6 +60,7 @@ export class SessionsService {
         }
         return await this.sessionsRepository.deleteSessionByDeviceId(deviceId)
     }
+
     async deleteActiveSessionWithoutCurrent(userId: string, iat: number): Promise<void> {
         const issueAt = new Date(iat * 1000).toISOString()
         await this.sessionsRepository.deleteSessionsWithoutCurrent(userId, issueAt)
